Abstracted the playlist sorting radio group into its own component
